Split out the letacy hook-based global task executor
